+++ src/extension.ts
@@ -7,37 +7,37 @@ export function activate(context: vscode.ExtensionContext) {
     let unquote = vscode.commands.registerCommand('extension.quotifyUnquoteStrings', () => unquoteStrings());
     let newline = vscode.commands.registerCommand('extension.quotifyNewlineStrings', () => newLineStrings());
     let commaNewLine = vscode.commands.registerCommand('extension.quotifyCommaNewlineStrings', () => commaNewLineStrings());
-    let arrayString = vscode.commands.registerCommand('extension.buildArrayString', () => buildArrayString());
+    // let arrayString = vscode.commands.registerCommand('extension.buildArrayString', () => buildArrayString());
     // let sqlSetString = vscode.commands.registerCommand('extension.buildSqlSetString', () => buildSqlSetString());
 
     context.subscriptions.push(quote);
     context.subscriptions.push(unquote);
     context.subscriptions.push(newline);
     context.subscriptions.push(commaNewLine);
-    context.subscriptions.push(arrayString);
+    // context.subscriptions.push(arrayString);
     // context.subscriptions.push(sqlSetString);
 }
 
 export function deactivate() {
 }
 
-function buildArrayString() {
-    let editor = vscode.window.activeTextEditor;
-    let range;
-    if (!editor.selection.isEmpty) {
-        range = editor.selection;
-    } else {
-        vscode.window.showErrorMessage("No text selected");
-        return;
-    }
+// function buildArrayString() {
+//     let editor = vscode.window.activeTextEditor;
+//     let range;
+//     if (!editor.selection.isEmpty) {
+//         range = editor.selection;
+//     } else {
+//         vscode.window.showErrorMessage("No text selected");
+//         return;
+//     }
 
-    let text = editor.document.getText(range);
-    let result = JSON.stringify(splitByCommaOrNewLine(text));
-    editor.edit((builder) => {
-        builder.replace(range, result);
-    });
-    editor.revealRange(range);
-}
+//     let text = editor.document.getText(range);
+//     let result = JSON.stringify(splitByCommaOrNewLine(text));
+//     editor.edit((builder) => {
+//         builder.replace(range, result);
+//     });
+//     editor.revealRange(range);
+// }
 
 // function buildSqlSetString() {
 //     let editor = vscode.window.activeTextEditor;