+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to the "quotify" extension will be documented in this file.
 
+## [0.0.4]
+- Prefix all commands with word 'Quotify'
+- Allow user to select their choice of quote to use during quotification process
+
 ## [0.0.3]
 - Comma newline delmited text
 - Have a sql list but need more robust testing around it and escaping single ticks